Package-Struktur der Tests verändert
