+++ src/main/java/dhbw/mwi/Auslandsemesterportal2016/rest/CreateSGLServlet.java
@@ -15,7 +15,7 @@ import javax.servlet.http.HttpServletResponse;
 
 import dhbw.mwi.Auslandsemesterportal2016.db.SQL_queries;
 import dhbw.mwi.Auslandsemesterportal2016.db.Util;
-import dhbw.mwi.Auslandsemesterportal2016.db.userAuthentification;
+import dhbw.mwi.Auslandsemesterportal2016.db.UserAuthentification;
 
 @WebServlet(name = "createSGLServlet", urlPatterns = { "/createSGL" })
 public class CreateSGLServlet extends HttpServlet {
@@ -25,7 +25,7 @@ public class CreateSGLServlet extends HttpServlet {
 		Util.addResponseHeaders(request, response);
 
 		PrintWriter out = response.getWriter();
-		int rolle = userAuthentification.isUserAuthentifiedByCookie(request);
+		int rolle = UserAuthentification.isUserAuthentifiedByCookie(request);
 
 		if (rolle != 1) {
 			response.sendError(401, "Rolle: " + rolle);
