+++ src/main/java/dhbw/mwi/Auslandsemesterportal2016/rest/GetCurrentActivityServlet.java
@@ -16,7 +16,7 @@ import com.google.gson.JsonObject;
 
 import dhbw.mwi.Auslandsemesterportal2016.db.SQL_queries;
 import dhbw.mwi.Auslandsemesterportal2016.db.Util;
-import dhbw.mwi.Auslandsemesterportal2016.db.userAuthentification;
+import dhbw.mwi.Auslandsemesterportal2016.db.UserAuthentification;
 
 @WebServlet(name = "GetCurrentActivityServlet", urlPatterns = { "/currentActivity" })
 public class GetCurrentActivityServlet extends HttpServlet {
@@ -25,7 +25,7 @@ public class GetCurrentActivityServlet extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
 		Util.addResponseHeaders(request,response);
 
-		int rolle = userAuthentification.isUserAuthentifiedByCookie(request);
+		int rolle = UserAuthentification.isUserAuthentifiedByCookie(request);
 
 		if (rolle < 1) {
 			response.sendError(401);
