+++ src/main/java/dhbw/mwi/Auslandsemesterportal2016/rest/GetEmailTextServlet.java
@@ -13,7 +13,7 @@ import org.camunda.bpm.engine.ProcessEngine;
 import org.camunda.bpm.engine.ProcessEngines;
 import org.camunda.bpm.engine.RuntimeService;
 
-import dhbw.mwi.Auslandsemesterportal2016.db.userAuthentification;
+import dhbw.mwi.Auslandsemesterportal2016.db.UserAuthentification;
 
 @WebServlet(name = "GetEmailTextServlet", urlPatterns = { "/getMailText" })
 public class GetEmailTextServlet extends HttpServlet {
@@ -22,7 +22,7 @@ public class GetEmailTextServlet extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
 		Util.addResponseHeaders(request, response);
 
-		int rolle = userAuthentification.isUserAuthentifiedByCookie(request);
+		int rolle = UserAuthentification.isUserAuthentifiedByCookie(request);
 
 		if (rolle != 1 && rolle != 2 && rolle != 4) {
 			response.sendError(401);
