+++ src/main/java/dhbw/mwi/Auslandsemesterportal2016/rest/GetOverviewServlet.java
@@ -13,7 +13,7 @@ import com.google.gson.JsonObject;
 
 import dhbw.mwi.Auslandsemesterportal2016.db.SQL_queries;
 import dhbw.mwi.Auslandsemesterportal2016.db.Util;
-import dhbw.mwi.Auslandsemesterportal2016.db.userAuthentification;
+import dhbw.mwi.Auslandsemesterportal2016.db.UserAuthentification;
 
 @WebServlet(name = "GetOverviewServlet", urlPatterns = { "/getOverview" })
 public class GetOverviewServlet extends HttpServlet {
@@ -22,7 +22,7 @@ public class GetOverviewServlet extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
 		Util.addResponseHeaders(request,response);
 
-		int rolle = userAuthentification.isUserAuthentifiedByCookie(request);
+		int rolle = UserAuthentification.isUserAuthentifiedByCookie(request);
 
 		if (rolle < 1) {
 			response.sendError(401);
