+++ src/main/java/dhbw/mwi/Auslandsemesterportal2016/rest/GetVariablesServlet.java
@@ -15,7 +15,7 @@ import org.camunda.bpm.engine.RuntimeService;
 import com.google.gson.JsonObject;
 
 import dhbw.mwi.Auslandsemesterportal2016.db.Util;
-import dhbw.mwi.Auslandsemesterportal2016.db.userAuthentification;
+import dhbw.mwi.Auslandsemesterportal2016.db.UserAuthentification;
 
 @WebServlet(name = "GetVariablesServlet", urlPatterns = { "/getVariables" })
 public class GetVariablesServlet extends HttpServlet {
@@ -23,7 +23,7 @@ public class GetVariablesServlet extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
 		Util.addResponseHeaders(request,response);
-		int rolle = userAuthentification.isUserAuthentifiedByCookie(request);
+		int rolle = UserAuthentification.isUserAuthentifiedByCookie(request);
 		if (rolle < 1) {
 			response.sendError(401);
 		} else {
