+++ src/main/java/dhbw/mwi/Auslandsemesterportal2016/rest/Logout.java
@@ -10,7 +10,7 @@ import javax.servlet.http.HttpServletResponse;
 
 import dhbw.mwi.Auslandsemesterportal2016.db.SQL_queries;
 import dhbw.mwi.Auslandsemesterportal2016.db.Util;
-import dhbw.mwi.Auslandsemesterportal2016.db.userAuthentification;
+import dhbw.mwi.Auslandsemesterportal2016.db.UserAuthentification;
 
 @WebServlet(urlPatterns = { "/logout" })
 public class Logout extends HttpServlet {
@@ -20,7 +20,7 @@ public class Logout extends HttpServlet {
 		Util.addResponseHeaders(request,response);
 
 		// ONLY VALID SESSIONS CAN LOGOUT
-		int rolle = userAuthentification.isUserAuthentifiedByCookie(request);
+		int rolle = UserAuthentification.isUserAuthentifiedByCookie(request);
 
 		if (rolle > 1) {
 			Cookie[] cookies = request.getCookies();
