+++ src/main/java/dhbw/mwi/Auslandsemesterportal2016/rest/UploadServlet.java
@@ -17,7 +17,7 @@ import org.camunda.bpm.engine.ProcessEngines;
 import org.camunda.bpm.engine.variable.Variables;
 import org.camunda.bpm.engine.variable.value.FileValue;
 
-import dhbw.mwi.Auslandsemesterportal2016.db.userAuthentification;
+import dhbw.mwi.Auslandsemesterportal2016.db.UserAuthentification;
 
 @WebServlet(name = "UploadServlet", urlPatterns = { "/upload" })
 @MultipartConfig(maxFileSize = 16177215) // 16MB
@@ -29,7 +29,7 @@ public class UploadServlet extends HttpServlet {
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
 		Util.addResponseHeaders(request,response);
 
-		int rolle = userAuthentification.isUserAuthentifiedByCookie(request);
+		int rolle = UserAuthentification.isUserAuthentifiedByCookie(request);
 
 		if (rolle < 1) {
 			response.sendError(401);
