+++ src/main/java/dhbw/mwi/Auslandsemesterportal2016/rest/UserUpdateServlet.java
@@ -13,7 +13,7 @@ import javax.servlet.http.HttpServletResponse;
 import dhbw.mwi.Auslandsemesterportal2016.Config;
 import dhbw.mwi.Auslandsemesterportal2016.db.SQL_queries;
 import dhbw.mwi.Auslandsemesterportal2016.db.Util;
-import dhbw.mwi.Auslandsemesterportal2016.db.userAuthentification;
+import dhbw.mwi.Auslandsemesterportal2016.db.UserAuthentification;
 import dhbw.mwi.Auslandsemesterportal2016.enums.ErrorEnum;
 import dhbw.mwi.Auslandsemesterportal2016.enums.SuccessEnum;
 
@@ -23,7 +23,7 @@ public class UserUpdateServlet extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
         Util.addResponseHeaders(request, response);
-        int rolle = userAuthentification.isUserAuthentifiedByCookie(request);
+        int rolle = UserAuthentification.isUserAuthentifiedByCookie(request);
         if (rolle != 1) {
             response.sendError(401, "Rolle: " + rolle);
         } else {
