+++ src/test/java/dhbw/mwi/Auslandsemesterportal2016/test/db/UserAuthentificationTest.java
@@ -1,0 +1,7 @@
+package dhbw.mwi.Auslandsemesterportal2016.test.db;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class UserAuthentificationTest {
+
+}
