+++ src/test/java/dhbw/mwi/Auslandsemesterportal2016/test/rest/CreateStudentServletTest.java
@@ -1,11 +1,11 @@
-package dhbw.mwi.Auslandsemesterportal2016.test;
+package dhbw.mwi.Auslandsemesterportal2016.test.rest;
 
 import dhbw.mwi.Auslandsemesterportal2016.db.SQL_queries;
-import dhbw.mwi.Auslandsemesterportal2016.db.userAuthentification;
+import dhbw.mwi.Auslandsemesterportal2016.db.UserAuthentification;
 import dhbw.mwi.Auslandsemesterportal2016.enums.ErrorEnum;
 import dhbw.mwi.Auslandsemesterportal2016.enums.SuccessEnum;
 import dhbw.mwi.Auslandsemesterportal2016.enums.TestEnum;
-import dhbw.mwi.Auslandsemesterportal2016.rest.CreateAAAServlet;
+import dhbw.mwi.Auslandsemesterportal2016.rest.CreateStudentServlet;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -25,11 +25,10 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
-class CreateAAAServletTest {
+class CreateStudentServletTest {
     // Initialization of necessary mock objects for mocking instance methods
     ResultSet resultSet = mock(ResultSet.class);
     HttpServletRequest request = mock(HttpServletRequest.class);
@@ -43,15 +42,14 @@ class CreateAAAServletTest {
     StringWriter stringWriter;
     PrintWriter writer;
     Cookie c1 = new Cookie("email", TestEnum.TESTEMAIL.toString());
-    Cookie c2 = new Cookie("sessionID", TestEnum.TESTSESSIONID.toString());
+    Cookie c2 = new Cookie("sessionID", "s1e5f2ge8gvs694g8vedsg");
     Cookie[] cookies = { c1, c2 };
-    CreateAAAServlet aaaServlet = new CreateAAAServlet();
+    CreateStudentServlet studentServlet = new CreateStudentServlet();
 
     @BeforeEach
     public void init() throws IOException, SQLException {
         // Define necessary mock objects for mocking static methods
         sql_queries = Mockito.mockStatic(SQL_queries.class);
-
         // Define necessary instances
         stringWriter = new StringWriter();
         writer = new PrintWriter(stringWriter);
@@ -68,8 +66,11 @@ class CreateAAAServletTest {
         when(request.getParameter("email")).thenReturn(TestEnum.TESTEMAIL.toString());
         when(request.getParameter("vorname")).thenReturn(TestEnum.TESTVNAME.toString());
         when(request.getParameter("nachname")).thenReturn(TestEnum.TESTNNAME.toString());
-        when(request.getParameter("tel")).thenReturn(TestEnum.TESTTELNR.toString());
-        when(request.getParameter("mobil")).thenReturn(TestEnum.TESTMOBILNR.toString());
+        when(request.getParameter("studgang")).thenReturn(TestEnum.TESTSTUGANG.toString());
+        when(request.getParameter("kurs")).thenReturn(TestEnum.TESTKURS.toString());
+        when(request.getParameter("matnr")).thenReturn(TestEnum.TESTMATRNR.toString());
+        when(request.getParameter("standort")).thenReturn(TestEnum.TESTSTANDORT.toString());
+        when(request.getRequestDispatcher(anyString())).thenReturn(requestDispatcher);
 
         // 1 = Admin
         when(resultSet.getInt(anyInt())).thenReturn(1);
@@ -91,14 +92,13 @@ class CreateAAAServletTest {
         sql_queries.when(() -> SQL_queries.userRegister(anyString(), anyString(), anyString(), anyString(), anyInt(),
                         anyString(), anyString(), anyString(), anyInt(), anyString(), anyString(), anyString(), anyString()))
                 .thenCallRealMethod();
-        when(request.getRequestDispatcher(anyString())).thenReturn(requestDispatcher);
 
         Mockito.doAnswer((Answer<Object>) invocation -> {
             writer.print(SuccessEnum.CREATEUSER);
             return null;
         }).when(requestDispatcher).forward(any(), any());
 
-        aaaServlet.doPost(request, response);
+        studentServlet.doPost(request, response);
 
         // get the value of stringWriter
         String result = stringWriter.toString().trim();
@@ -108,10 +108,10 @@ class CreateAAAServletTest {
     @Test
     void doPostUnauthorizedRole() throws IOException {
         int rolle = 2;
-        MockedStatic<userAuthentification> userAuthentificationMock = Mockito.mockStatic(userAuthentification.class);
-        userAuthentificationMock.when(() -> userAuthentification.isUserAuthentifiedByCookie(request)).thenReturn(rolle);
+        MockedStatic<UserAuthentification> userAuthentificationMock = Mockito.mockStatic(UserAuthentification.class);
+        userAuthentificationMock.when(() -> UserAuthentification.isUserAuthentifiedByCookie(request)).thenReturn(rolle);
 
-        aaaServlet.doPost(request, response);
+        new CreateStudentServlet().doPost(request, response);
 
         verify(response, times(1)).sendError(401, "Rolle: " + rolle);
 
@@ -122,7 +122,7 @@ class CreateAAAServletTest {
     void doPostEmailAlreadyUsed() throws IOException {
         sql_queries.when(() -> SQL_queries.isEmailUsed(any())).thenReturn(true);
 
-        aaaServlet.doPost(request, response);
+        studentServlet.doPost(request, response);
 
         String result = stringWriter.toString().trim();
         assertEquals(ErrorEnum.MAILERROR.toString(), result);
@@ -134,22 +134,9 @@ class CreateAAAServletTest {
                         anyString(), anyString(), anyString(), anyInt(), anyString(), anyString(), anyString(), anyString()))
                 .thenReturn(0);
 
-        aaaServlet.doPost(request, response);
+        studentServlet.doPost(request, response);
 
         String result = stringWriter.toString().trim();
         assertEquals(ErrorEnum.USERREGISTER.toString(), result);
-    }
-
-    @Test
-    void doPostInterruptedByException() throws IOException {
-        sql_queries.when(() -> SQL_queries.userRegister(anyString(), anyString(), anyString(), anyString(), anyInt(),
-                        anyString(), anyString(), anyString(), anyInt(), anyString(), anyString(), anyString(), anyString()))
-                .thenCallRealMethod();
-        // throw any Exception in try-Block
-        when(request.getRequestDispatcher(anyString())).thenThrow(RuntimeException.class);
-
-        assertThrows(RuntimeException.class, () -> aaaServlet.doPost(request, response));
-        //TODO check why e.getMessage() is always null
-        verify(response, times(1)).sendError(500, ErrorEnum.USERCREATE + "null");
     }
 }
