+++ src/test/java/dhbw/mwi/Auslandsemesterportal2016/test/rest/LoginServletTest.java
@@ -1,4 +1,4 @@
-package dhbw.mwi.Auslandsemesterportal2016.test;
+package dhbw.mwi.Auslandsemesterportal2016.test.rest;
 
 import dhbw.mwi.Auslandsemesterportal2016.db.SQL_queries;
 import dhbw.mwi.Auslandsemesterportal2016.enums.TestEnum;
